# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "gestioneventos",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/Eventos": {
            "get": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Listar todos los eventos",
                "description": "Obtener una lista de todos los eventos. Se puede aplicar un filtro por tipo de evento.",
                "operationId": "getEventos",
                "parameters": [
                    {
                        "name": "tipo_evento",
                        "in": "query",
                        "description": "Filtrar eventos por tipo",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de eventos obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Eventos": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Evento"
                                            }
                                        },
                                        "status": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontró ningún evento con ese tipo.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Crear un nuevo evento",
                "description": "Registrar un nuevo evento en el sistema.",
                "operationId": "createEvento",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre_evento": {
                                        "type": "string",
                                        "example": "Rodrigo y Daniela"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Boda Rustica"
                                    },
                                    "tipo_evento": {
                                        "type": "string",
                                        "example": "Matrimonio"
                                    },
                                    "fecha_conclusion": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-12-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Evento creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Eventos": {
                                            "$ref": "#/components/schemas/Evento"
                                        },
                                        "status": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la validación de los datos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        },
                                        "status": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Evento": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nombre_evento": {
                        "type": "string"
                    },
                    "descripcion": {
                        "type": "string"
                    },
                    "tipo_evento": {
                        "type": "string"
                    },
                    "fecha_conclusion": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}